+++ utils/event-apis.js
@@ -3,6 +3,10 @@ const API_QUERY_PARAM = 'featuredCards';
 
 const pageDataCache = {};
 
+export function getEventId() {
+  return window.bm8tr.get('eventData')?.arbitrary?.[0]?.value?.split('|')?.[1];
+}
+
 export function flattenObject(obj, parentKey = '', result = {}) {
   Object.keys(obj).forEach((key) => {
     const value = obj[key];
@@ -80,7 +84,8 @@ export async function getProfile() {
   return {};
 }
 
-export async function getAttendeeData(email, eventId) {
+// TODO: remove placeholder event ID
+export async function getAttendeeData(email, eventId = '458926431') {
   const myHeaders = new Headers();
   myHeaders.append('x-api-key', 'CCHomeWeb1');
 
@@ -90,15 +95,44 @@ export async function getAttendeeData(email, eventId) {
     redirect: 'follow',
   };
 
-  // TODO: use real event ID when ready
-  const data = await fetch(`https://cchome-stage.adobe.io/lod/v1/events/st-458926431/attendees/${email}`, requestOptions)
-  // const data = await fetch(`https://cchome-stage.adobe.io/lod/v1/events/st-${eventId}/attendees/${email}`, requestOptions)
+  const data = await fetch(`https://cchome-stage.adobe.io/lod/v1/events/st-${eventId}/attendees/${email}`, requestOptions)
     .then((response) => response.json())
     .then((result) => result)
     .catch((error) => console.error(error));
 
   console.log('Fetched attendee data:', data);
   return data;
+}
+
+export async function submitToSplashThat(payload) {
+  const myHeaders = new Headers();
+  myHeaders.append('x-api-key', 'CCHomeWeb1');
+  myHeaders.append('Content-Type', 'application/json');
+
+  const raw = JSON.stringify(payload);
+
+  const requestOptions = {
+    method: 'POST',
+    headers: myHeaders,
+    body: raw,
+    redirect: 'follow',
+  };
+
+  const eventId = getEventId();
+
+  if (!eventId) return false;
+  // TODO: use real event ID when ready
+  const resp = await fetch('https://cchome-stage.adobe.io/lod/v1/events/st-458926431/attendees', requestOptions).then((response) => response);
+  // const resp = await fetch(`https://cchome-stage.adobe.io/lod/v1/events/st-${eventId}/attendees`, requestOptions).then((response) => response);
+
+  console.log('Submitted registration to SplashThat:', payload);
+  resp.json().then((json) => {
+    console.log('Event Service Layer response:', json);
+  });
+
+  if (!resp.ok) return false;
+
+  return payload;
 }
 
 function lazyCaptureProfile() {
@@ -126,10 +160,6 @@ function lazyCaptureProfile() {
       attempCounter += 1;
     }
   }, 1000);
-}
-
-export function getEventId() {
-  return window.bm8tr.get('eventData')?.arbitrary?.[0]?.value?.split('|')?.[1];
 }
 
 export default async function fetchPageData(hash, lazyLoadProfile = false) {
